[E2E] Support testing remote-server template (#1138)
* Fix missing RBAC for remote servers

* Remove hosted from the default testing configuration

* Add e2e testing for the remote server

* Address review comments

* Add resource-policy: keep annotation on K0sWorkerConfig

To address https://github.com/k0smotron/k0smotron/issues/942

* Extend the list of namespaces to collect logs from

* Stream the command outputs to GinkgoWriter

Also, remove several unused functions

* Change kubevirt vm source to ubuntu and decrease the storage request

* Fix remote testing cluster cleanup

* [e2e] Add missing nil checks to deletion functions

* Add comment with annotation description to remote-server K0sWorkerConfig

+++ test/e2e/provider_aws_test.go
@@ -81,8 +81,10 @@ var _ = Describe("AWS Templates", Label("provider:cloud", "provider:aws"), Order
 			By("deleting resources")
 			deleteFuncs := append(hostedDeleteFuncs, append(standaloneDeleteFuncs, kubeconfigDeleteFuncs...)...)
 			for _, deleteFunc := range deleteFuncs {
-				err := deleteFunc()
-				Expect(err).NotTo(HaveOccurred())
+				if deleteFunc != nil {
+					err := deleteFunc()
+					Expect(err).NotTo(HaveOccurred())
+				}
 			}
 		}
 	})
